Änderung des Parameter auf Long

+++ inoUnitReihe/ClsFunktion.vb
@@ -1,6 +1,6 @@
 ﻿Public Class ClsFunktion
-    Public Function Reihe(Eingabe As Short) As Short
-        Dim Ergebnis As Short
+    Public Function Reihe(Eingabe As Long) As Long
+        Dim Ergebnis As Long
 
         For I As Short = 1 To Eingabe
             Ergebnis += I
